new: add exceptions for file handler
- update tests for plugin service

+++ src/main/kotlin/com/jtm/minecraft/core/usecase/file/FileHandler.kt
@@ -1,6 +1,6 @@
 package com.jtm.minecraft.core.usecase.file
 
-import lombok.extern.slf4j.Slf4j
+import com.jtm.minecraft.core.domain.exceptions.FileNotFound
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.http.codec.multipart.FilePart
@@ -8,28 +8,31 @@ import org.springframework.stereotype.Component
 import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
 import java.io.File
+
 @Component
 class FileHandler {
 
-    val logger = LoggerFactory.getLogger(FileHandler::class.java)
+    private val logger = LoggerFactory.getLogger(FileHandler::class.java)
 
-    @Value("\${disk.path:path}")
+    @Value("\${disk.path:/disk}")
     lateinit var disk: String
 
     fun save(path: String, filePart: FilePart): Mono<Void> {
-        val folder = File(path)
+        val folder = File(disk + path)
         if (!folder.exists()) if (folder.mkdirs()) logger.info("Created directories at: $path")
-        val file = File(path, filePart.filename())
+        val file = File(disk + path, filePart.filename())
         return filePart.transferTo(file)
     }
 
     fun fetch(path: String): Mono<File> {
-        val file = File(path)
+        val file = File(disk + path)
+        if (!file.exists()) return Mono.error { FileNotFound() }
         return Mono.just(file)
     }
 
     fun delete(path: String): Mono<File> {
-        val file = File(path)
+        val file = File(disk + path)
+        if (!file.exists()) return Mono.error { FileNotFound() }
         return Mono.just(file)
             .map {
                 it.delete()
@@ -38,7 +41,7 @@ class FileHandler {
     }
 
     fun listFiles(path: String): Flux<File> {
-        val folder = File(path)
+        val folder = File(disk + path)
         return Flux.fromArray(folder.listFiles())
     }
 }
